% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broadcast.R
\name{broadcast}
\alias{broadcast}
\title{broadcasting a function to all other rows}
\usage{
broadcast(nest_df, f, output_type = "list")
}
\arguments{
\item{nest_df}{The nested data frame to be broadcasted.
Must include a column as dataframe type with name 'data'.}

\item{f}{The function to be broadcasted.}

\item{output_type}{The expected output column type. The default is a list.
Options: 'list'; 'double'}
}
\description{
This function will return a dataframe with last item being the result of broadcasting.
}
