% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broadcast.R
\name{broadcast_group}
\alias{broadcast_group}
\title{broadcasting a function with grouping variables to all other rows}
\usage{
broadcast_group(nest_df, f)
}
\arguments{
\item{nest_df}{The nested data frame to be broadcasted.
Must include a column as dataframe type with name 'data'.}

\item{f}{The function to be broadcasted.}
}
\description{
This function will return a dataframe with last item (named \code{output})
being the result of broadcasting.
}
