% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezPurrr-package.R
\docType{package}
\name{ezPurrr-package}
\alias{ezPurrr}
\alias{ezPurrr-package}
\title{ezPurrr: Make Purrr Easier To Use.}
\description{
This is a facilitating package for Purrr with easy sampling and apply functions to all rows for a nested dataset.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://wanjiag.github.io/ezPurrr/}
}

}
\author{
\strong{Maintainer}: Wanjia Guo \email{wanjiag@uoregon.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
