% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesubset.R
\name{sample_data}
\alias{sample_data}
\title{Sampling one data frame from a nested dataframe. The sampling could be done
either randomly or a specific row could be indicated.}
\usage{
sample_data(nest_df, index)
}
\arguments{
\item{nest_df}{The nested dataframe to be sampled from.
The right-most column must be a dataframe type with name 'data'.}

\item{index}{Optional. The index of the dataframe to be sampled from. If no index is
provided, a random subset will be returned.}
}
\description{
This function will return a dataframe from the list of data frames from the
input nested dataframe.
}
