% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesubset.R
\name{sample_group}
\alias{sample_group}
\title{Sampling one grouping variable from a nested dataframe. The sampling could be done
either randomly or a specific row could be indicated.}
\usage{
sample_group(nest_df, index)
}
\arguments{
\item{nest_df}{The nested dataframe to be sampled from.
The right-most column must be a dataframe type with name 'data'.}

\item{index}{Optional. The index of the dataframe to be sampled from. If no index is
provided, a random subset will be returned.}
}
\description{
This function will return a dataframe of the grouping
variable from the input nested dataframe.
}
