% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesubset.R
\name{sampling_check}
\alias{sampling_check}
\title{Sanity check for input dataframe, then return the index of the row to be sampled from.}
\usage{
sampling_check(nest_df, index)
}
\arguments{
\item{nest_df}{The nested dataframe to be sampled from.
The right-most column must be a dataframe type with name 'data'.}

\item{index}{Optional. The index of the dataframe to be sampled from. If no index is
provided, a random subset will be returned.}
}
\description{
This function will return an index to be sampled from the input dataframe.
}
\keyword{Internal}
