% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezPurrr-package.R
\docType{package}
\name{ezPurrr-package}
\alias{ezPurrr}
\alias{ezPurrr-package}
\title{ezPurrr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Wanjia Guo \email{wanjiag@uoregon.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
