% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkgroup.R
\name{samplesubset}
\alias{samplesubset}
\title{sampling a subset of data from a nested data frame}
\usage{
samplesubset(nest_df, index)
}
\arguments{
\item{nest_df}{The nested data frame to be subset from}

\item{index}{The index of the data set to be returned from. If no index is
provided, a random subset will be returned. need to indicate this is optional}
}
\description{
This function will return a list with the first item being the grouping
variable of the data frame, and second item being the subsetted datafram.
}
